BREAKING: use a minor mode to set up pulsar
This supersedes the old design which has a built-in assumption that the
"pulse line" effect should always be global.  Now the user can have a
global style or a buffer-local one.  They may also disable pulsar on
demand.

Note that pulsar still only triggers its effect for entries in the user
option 'pulsar-pulse-functions'.

Thanks to Rudolf Adamkovič for the feedback in issue 9:
<https://gitlab.com/protesilaos/pulsar/-/issues/9>.