Relax test for last pulsar command
This matches commands that the user may define such as "my/pulsar-WORD",
whereas before it was limited to Pulsar's own prefix.

I understand that 'string-match-p' is more likely to return some false
positives, though I feel the likelihood is low.

See commit 2c2600c for the rationale of this condition.