DEPRECATE pulsar-pulse-on-window-change
It adds complexity to an otherwise simple code base.  I noticed this
with commits 1848ba0 and 2c2600c.  There may be more cases to consider,
which makes me think this feature is not worth having.

This change may also be related to what djl reports in issue 3 on the
GitHub mirror about not pulsing in the minibuffer:
<https://github.com/protesilaos/pulsar/issues/3>.