Read real-this-command if necessary
The need for this arrangement became clear after trying to make the
scroll motions of the 'evil' package work with pulsar-pulse-functions.

Thanks to Duy Nguyen for reporting the issue on the mailing list.  I
helped find the source of the apparent problem in the 'evil' code base
and then Tom Dalziel explained why 'evil' does things this way,
suggesting 'real-this-command' as a possibility:

- <https://lists.sr.ht/~protesilaos/pulsar/%3C89566F5C-25AD-4281-94CB-031FE8878119%40gmail.com%3E>
- <https://lists.sr.ht/~protesilaos/pulsar/%3C87pmgy3vzq.fsf%40protesilaos.com%3E>
- <https://github.com/emacs-evil/evil/issues/1659>