Define a helper function to get the line boundaries
We want to use our other helper functions instead of using
line-beginning-position and line-end-position. Those are okay, but
they are not exactly what we are looking for, as we also want to
extend the highlight to the edge of the window. For the beginning
position, we have to consider the context as, for example, in the
minibuffer we want something else as the beginning than in a regular
buffer.

The pulsar--start and pulsar--end can probably be simplified as well.
I will need to review them at some point.