Remove needless check for the minor mode in post command pulse and make the function public
I think it is better to have public functions added to hooks. Those
are supposed to be exposed to the user, after all.

The check for the minor mode is not needed in the context of this
function. The minor mode is responsible for handling it via the
relevant hooks.