Be consistent about pulsing to avoid unwanted static highlights
Thanks to Koloszár Gergely, Mehrad Mahmoudian, and Nikolaos
Bezirgiannis for reporting that the Emacs daemon would produce static
highlights instead of pulsing them in and out of effect. This was done
in issue 32: <https://github.com/protesilaos/pulsar/issues/32>.

I do not know why this was only happening with the Emacs daemon, but
this change is better regardless.